pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "levitate-trading-stack"
version = "0.1.0"
description = "Config-driven, event-sourced trading stack scaffold (market feed, strategy, execution, analytics, UI)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyarrow>=14.0",
  "pydantic>=2.0",
  "plotly>=5.18",
  "streamlit>=1.30",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
levitate = "runner.cli:main"

[tool.setuptools]
# prevents setuptools from trying to guess / auto-include everything
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["common*", "runner*", "market_feed*", "execution*", "strategy*", "analytics*", "preprocess*"]
exclude = ["configs*", "ui*", "runs*"]
